test: guard pinned imports through an unlocked multi-file bun run
